added new features: 1. utilises folder that does the http request to the app.py with jinja 2. created a scratch for the landing using jinja ( it does not see the static folder, so for now decided that we can go without it and maybe change this later) 3. crated a basic landing page, the images should be optimized, the last circle should be deleted and the text should be added. 4. created a new folder documentation_styles where are all the templates for the documentation.
