+++ documentation_styles/templates/index.html
@@ -1,0 +1,41 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>Greeting</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
+</head>
+<style>
+
+    .code{
+        text-align: left;
+        background-color: #f5f8fa !important;
+        color: #0c5176 !important;
+    }
+</style>
+<body>
+    <div class="card ">
+        <h1 class="card-title">{{module_name}}</h1>
+
+        <div class="card-body" style="width:50rem;">
+            <p class="card-text">
+            {{ paragraph }}
+            </p>
+
+            <pre class="card code mb-0" style="width:max-content; padding-right: 20px;">
+                <code>
+                        {{code}}
+                </code>
+            </pre>
+            <pre><code>
+                &lt;p&gt;Sample text here...&lt;/p&gt;
+                &lt;p&gt;And another line of sample text here...&lt;/p&gt;
+            </code></pre>
+
+
+        </div>
+
+    </div>
+
+</body>
+</html>
