+++ landing/landing_simple/index.html
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
+    <link rel="stylesheet" href="style.css">
+`
+</head>
+<body class="jumbotron">
+<div class="header container">
+    <div class="row">
+        <div class="col-sm">
+            <img src="../static/assets-rubicon/rotating-small-logo.gif" alt="">
+        </div>
+        <div class="col-lg-5 mt-5">
+            <!-- <img src="../static/assets-rubicon/rotating-large-cube.gif" alt=""> -->
+            <img class="rotating" src="../static/assets-rubicon/cube-frame.gif" alt="">
+        </div>
+        <div class="col-sm-offset-2 ml-5 pl-5">
+            <img src="../static/assets-rubicon/button-download.png" alt="">
+        </div>
+    </div>
+    <div class="row main-title">
+         <img src="../static/assets-rubicon/main-title.png" alt="" style="width: 90%">
+    </div>
+</div>
+<div class="container about">
+        <div class="row top-buffer"></div>
+        <div class="row">
+            <div class="col">
+                <img src="../static/assets-rubicon/write-code.png" alt="" style="width: 80%">
+            </div>
+            <div class="col-lg-offset-2">
+
+                <img src="../static/assets-rubicon/Rectangle%2019.png" alt="">
+
+            </div>
+        </div>
+        <div class="row top-buffer-xs"></div>
+        <div class="row">
+            <div class="col-lg-offset-2">
+                <img src="../static/assets-rubicon/Rectangle%2019.png" alt="">
+            </div>
+            <div class="col">
+                <img src="../static/assets-rubicon/save-time.png" alt="" style="width: 95%">
+
+            </div>
+        </div>
+
+    </div>
+</div>
+<div class="cube-container" style="position: relative;">
+    <!-- <img class="rotating" src="../static/assets-rubicon/rotating-large-cube.gif" alt=""> -->
+    <img class="rotating" src="../static/assets-rubicon/cube-frame.gif" alt="">
+    <img src="../static/assets-rubicon/button-download.png" alt="" style="width: 20%;">
+</div>
+</body>
+</html>
