+++ landing/landing_simple/style.css
@@ -1,0 +1,74 @@
+
+.jumbotron{
+ background-image: url('../static/assets-rubicon/background.png');
+ background-size: cover;
+ background-repeat: no-repeat;
+    background-position: center;
+    height: 4243px;
+    width: 100%;
+}
+
+.top-buffer {
+    margin-top:500px;
+}
+
+.top-buffer-xs {
+    margin-top:400px;
+}
+
+.cube-container {
+    position: relative;
+    background: url(../static/assets-rubicon/spinning-circle.gif) no-repeat;
+    width: 1001px;
+    height: 1000px;
+    background-size: cover;
+    display: flex;
+    flex-direction: column;
+    margin: auto;
+    justify-content: center;
+    align-items: center;
+}
+
+.center{
+    top: 50%;
+    left: 50%;
+    margin-right: -50%;
+    transform: translate(-50%, -50%)
+}
+
+
+@-webkit-keyframes rotating /* Safari and Chrome */ {
+    from {
+      -webkit-transform: rotate(0deg);
+      -o-transform: rotate(0deg);
+      transform: rotate(0deg);
+    }
+    to {
+      -webkit-transform: rotate(360deg);
+      -o-transform: rotate(360deg);
+      transform: rotate(360deg);
+    }
+  }
+  @keyframes rotating {
+    from {
+      -ms-transform: rotate(0deg);
+      -moz-transform: rotate(0deg);
+      -webkit-transform: rotate(0deg);
+      -o-transform: rotate(0deg);
+      transform: rotate(0deg);
+    }
+    to {
+      -ms-transform: rotate(360deg);
+      -moz-transform: rotate(360deg);
+      -webkit-transform: rotate(360deg);
+      -o-transform: rotate(360deg);
+      transform: rotate(360deg);
+    }
+  }
+  .rotating {
+    -webkit-animation: rotating 5s linear infinite;
+    -moz-animation: rotating 5s linear infinite;
+    -ms-animation: rotating 5s linear infinite;
+    -o-animation: rotating 5s linear infinite;
+    animation: rotating 5s linear infinite;
+  }
