+++ landing/templates/landing.html
@@ -1,0 +1,11 @@
+<!--https://medium.com/@bmcilhen/the-wonderful-world-of-templating-with-pythons-flask-framework-be7d39b8c4ae-->
+
+{% extends 'main_layout.html' %}
+   {% block body %}
+       <div class="jumbotron text-center">
+          <h1>Welcome To FlaskApp</h1>
+             <p class="lead">Lorem Ipsum I sispem Blah dy bloo</p>
+             <a href="/register" class="btn btn-primary btn-lg">Register</a>
+      </div>
+   {% endblock %}
+
