+++ landing/templates/main_layout.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+  <html>
+      <head>
+         <meta charset="utf-8">
+            <title>Hola</title>
+            <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+            <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
+
+      </head>
+      <body class="body" >
+{#        {% include 'navbar.html' %}#}
+        <div class="container">
+          {% block body %}{% endblock %}
+         </div>
+         <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+      </body>
+  </html>
