+++ landing/templates/navbar.html
@@ -1,0 +1,1 @@
+<h1>This tis navbar</h1>
