--- templates/index2.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>{{browser_title}}</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-giJF6kkoqNQ00vy+HMDP7azOuL0xtbfIcaT9wjKHr8RbDVddVHyTfAAsrekwKmP1" crossorigin="anonymous">
-</head>
-    <body>
-    <!--highlighted text,
-    inspect the (css),
-    images,
-    links
-    tables
-    -->
-        <div class="container">
-                  <h1>{{main_title}}</h1>
-                  <img class="img-fluid" alt="Responsive image" src={{image_about}} >
-                  <h2>{{sub_title}}</h2>
-                  <p>{{paragraph}}</p>
-            </div>
-
-            <div class="container">
-                <h2>Table of Contents</h2>
-                    <ul>
-                        {%- for line in bullet_point %}
-                         <a href="{{line.url}}" ><li >{{ line.name}}</li></a>
-                        {% endfor %}
-                    </ul>
-            </div>
-
-            <div class="container">
-                <table id="table" class="table-info table table-striped">
-                    {%- for name in table_title %}
-                         <tr>
-                             <th>{{name[0]}}</th>
-                             <th>{{name[1]}}</th>
-                         </tr>
-                    {% endfor %}
-                    {%- for line in table_list %}
-                         <tr>
-                             <td>{{line.name}}</td>
-                             <td>{{line.status}}</td>
-                         </tr>
-                    {% endfor %}
-                </table>
-            </div>
-
-            <div class="container">
-                {{paragraph2}}
-            </div>
-        </div>
-    </body>
-</html>
